Split functionality into separate package
